--- src/components/App.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-import Menu from "./menu/Menu";
-import './App.scss'
-
-const App: React.FC = () => {
-	return (
-		<div>
-			<Menu />
-		</div>
-	);
-};
-
-export default App;
