+++ src/components/NavBar/index.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+
+import Search from "../Search";
+import "./index.scss";
+
+const NavBar: React.FC = () => {
+  return (
+    <div className="menu">
+      <nav >
+        <ul className="left-nav">
+          <li className="menu-link">ACCESSORIES</li>
+          <li className="menu-link">GROCERIES</li>
+          <li className="menu-link">APPAREL</li>
+        </ul>
+        <ul className="right-nav">
+          <li className="search-box">
+            <Search />
+          </li>
+          <li className="menu-link">CART</li>
+          <li className="menu-link">PROFILE</li>
+        </ul>
+      </nav>
+    </div>
+  );
+};
+
+export default NavBar;
