+++ src/components/Search/index.scss
