+++ src/components/index.tsx
@@ -1,0 +1,3 @@
+import NavBar from "./NavBar";
+
+export { NavBar };
