+++ src/index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html>
-	<head lang="en">
-		<meta charset="utf-8" />
-		<title>Seleet</title>
-	</head>
-	<body>
-		<div id="root"></div>
-	</body>
+  <head lang="en">
+    <meta charset="utf-8" />
+    <title>Seleet</title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
 </html>
