+++ src/views/Home/index.scss
@@ -1,0 +1,7 @@
+main {
+  min-height: 100vh;
+
+  img {
+    width: 100%;
+  }
+}
