+++ src/views/Home/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+import "./index.scss";
+
+const Home: React.FC = () => {
+  return (
+    <main>
+      <img src="./src/images/summer.jpg" alt="" />
+    </main>
+  );
+};
+
+export default Home;
